Turn mhunter into a module
I moved the ArgumentParser section into the if name == main section so
that I can import mhunter as a module.

I started to write the data_grabber, which uses mhunter to read BOM
data, and saves it into my database. Eventually I will run this file
regularly, perhaps as a CRON job.